Fixed issues with the findsentimentScore() method

+++ project/src/main/java/com/google/sps/servlets/Article.java
@@ -30,9 +30,9 @@ public class Article {
   private Date pubDate;
   private String description;
   private String source;
-  private Integer sentiment;
+  private Float sentiment;
 
-  public Article(String title, String link, Date pubDate, String description, String source, Integer sentiment) {
+  public Article(String title, String link, Date pubDate, String description, String source, Float sentiment) {
     this.title = title;
     this.link = link;
     this.pubDate = pubDate;
@@ -65,7 +65,7 @@ public class Article {
     return String.format("%s %s %s %s %s", title, link, pubDate.toString(), description, source);
   }
 
-  public Integer getSentiment() {
+  public Float getSentiment() {
     return sentiment;
   }
 }
